requirements.txt
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
deep-translator==1.11.4
distro==1.9.0
faiss-cpu==1.9.0.post1
filelock==3.16.1
fonttools==4.55.3
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
googletrans==4.0.0rc1
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.12.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
hyperframe==5.2.0
idna==2.10
InstructorEmbedding==1.0.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.3.12
langchain-community==0.3.12
langchain-core==0.3.25
langchain-text-splitters==0.3.3
langsmith==0.2.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.18.4
networkx==3.4.2
numpy==2.2.0
openai==1.58.1
opencv-python-headless==4.10.0.84
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.0.0
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.1
pyarrow==18.1.0
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.25.1
pyparsing==3.2.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.22.3
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.6.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wordcloud==1.9.4
yarl==1.18.3
